Updated and better component Changed entire toolchain to create-react-library Updated entire code base
