+++ src/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "env": {
+    "jest": true
+  }
+}
