--- src/apollo/index.js
@@ -1,14 +1,0 @@
-// The Grandeur Cloud React SDK
-// This is where the revolution is gonna happen
-// It will act as an interface to JS SDK and will 
-// provide interface so that developers can use
-// Grandeur Cloud in their React Applications
-
-// Import Component and Wrapper
-import Apollo from './apollo.component';
-import withApollo from './apollo.wrapper';
-
-
-// Export
-export default Apollo;
-export {withApollo};
